bitclout-demo/requirements.txt
Flask==1.1.2
pymgclient==1.0.0